requirements.txt
tensorflow-cpu==2.8.0
absl-py
aiohttp
aiosignal
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout
attrs
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
brotlipy==0.7.0
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm==0.1.3
contourpy==1.0.7
cryptography
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
docopt==0.6.2
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.3
flatbuffers
fonttools==4.39.3
fqdn==1.5.1
frozenlist
gast
gensim
google-auth
google-auth-oauthlib
google-pasta
grpcio
h11==0.14.0
h5py
idna
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras
Keras-Preprocessing
kiwisolver==1.4.4
libclang==16.0.0
Markdown
MarkupSafe
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
motor==3.1.2
multidict
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy
oauthlib
opencv-python==4.7.0.72
opt-einsum
packaging
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.2.0
pooch
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19
psutil==5.9.5
pure-eval==0.2.2
pyasn1
pyasn1-modules==0.2.8
pycparser
pydantic==1.10.7
Pygments==2.15.1
PyJWT
pymongo==4.3.3
pyOpenSSL
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex
requests
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa
scikit-learn
scipy==1.10.1
Send2Trash==1.8.0
six
smart-open
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.26.1
termcolor
terminado==0.17.1
threadpoolctl
tinycss2==1.2.1
tornado==6.3.1
tqdm
traitlets==5.9.0
typing_extensions
tzdata==2023.3
uri-template==1.2.0
urllib3
uvicorn==0.21.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug
widgetsnbextension==4.0.7
win-inet-pton
wrapt
yarg==0.1.9
yarl
nltk